+++ gazebo/math/Box_TEST.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2012-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,11 +18,10 @@
 #include <gtest/gtest.h>
 
 #include "gazebo/math/Box.hh"
-#include "test/util.hh"
 
 using namespace gazebo;
 
-class BoxTest : public gazebo::testing::AutoLogFixture { };
+class BoxTest : public ::testing::Test { };
 
 class ExampleBox : public ::testing::Test
 {