+++ gazebo/sensors/RFIDSensor.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2012-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,29 +14,20 @@
  * limitations under the License.
  *
 */
-/* Desc: RFID Sensor
- * Author: Jonas Mellin & Zakiruz Zaman
- * Date: 6th December 2011
- */
-
 #ifdef _WIN32
   // Ensure that Winsock2.h is included before Windows.h, which can get
   // pulled in by anybody (e.g., Boost).
   #include <Winsock2.h>
 #endif
 
+#include "gazebo/msgs/msgs.hh"
+#include "gazebo/transport/transport.hh"
 #include "gazebo/physics/World.hh"
 #include "gazebo/physics/Entity.hh"
 
-#include "gazebo/common/Exception.hh"
-
-#include "gazebo/transport/Node.hh"
-#include "gazebo/transport/Publisher.hh"
-
-#include "gazebo/msgs/msgs.hh"
-
 #include "gazebo/sensors/RFIDTag.hh"
 #include "gazebo/sensors/SensorFactory.hh"
+#include "gazebo/sensors/RFIDSensorPrivate.hh"
 #include "gazebo/sensors/RFIDSensor.hh"
 
 using namespace gazebo;
@@ -46,7 +37,8 @@ GZ_REGISTER_STATIC_SENSOR("rfid", RFIDSensor)
 
 /////////////////////////////////////////////////
 RFIDSensor::RFIDSensor()
-  : Sensor(sensors::OTHER)
+: Sensor(sensors::OTHER),
+  dataPtr(new RFIDSensorPrivate)
 {
   this->active = false;
 }
@@ -72,11 +64,11 @@ void RFIDSensor::Load(const std::string &_worldName)
 
   if (this->sdf->GetElement("topic"))
   {
-    this->scanPub = this->node->Advertise<msgs::Pose>(
+    this->dataPtr->scanPub = this->node->Advertise<msgs::Pose>(
         this->sdf->GetElement("topic")->Get<std::string>());
   }
 
-  this->entity = this->world->GetEntity(this->parentName);
+  this->dataPtr->entity = this->world->GetEntity(this->ParentName());
 
   // this->sdf->PrintDescription("something");
   /*std::cout << " setup ray" << std::endl;
@@ -132,11 +124,11 @@ bool RFIDSensor::UpdateImpl(const bool /*_force*/)
   this->EvaluateTags();
   this->lastMeasurementTime = this->world->GetSimTime();
 
-  if (this->scanPub)
+  if (this->dataPtr->scanPub)
   {
     msgs::Pose msg;
-    msgs::Set(&msg, this->entity->GetWorldPose().Ign());
-    this->scanPub->Publish(msg);
+    msgs::Set(&msg, this->dataPtr->entity->GetWorldPose().Ign());
+    this->dataPtr->scanPub->Publish(msg);
   }
 
   return true;
@@ -148,7 +140,7 @@ void RFIDSensor::EvaluateTags()
   std::vector<RFIDTag*>::const_iterator ci;
 
   // iterate through the tags contained given rfid tag manager
-  for (ci = this->tags.begin(); ci != this->tags.end(); ++ci)
+  for (ci = this->dataPtr->tags.begin(); ci != this->dataPtr->tags.end(); ++ci)
   {
     ignition::math::Pose3d pos = (*ci)->TagPose();
     // std::cout << "link: " << tagModelPtr->GetName() << std::endl;
@@ -164,7 +156,7 @@ bool RFIDSensor::CheckTagRange(const ignition::math::Pose3d &_pose)
 {
   // copy sensor vector pos into a temp var
   ignition::math::Vector3d v;
-  v = _pose.Pos() - this->entity->GetWorldPose().Ign().Pos();
+  v = _pose.Pos() - this->dataPtr->entity->GetWorldPose().Ign().Pos();
 
   // std::cout << v.GetLength() << std::endl;
 
@@ -181,5 +173,5 @@ bool RFIDSensor::CheckTagRange(const ignition::math::Pose3d &_pose)
 //////////////////////////////////////////////////
 void RFIDSensor::AddTag(RFIDTag *_tag)
 {
-  this->tags.push_back(_tag);
+  this->dataPtr->tags.push_back(_tag);
 }