+++ test/integration/dem.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2012-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -68,7 +68,7 @@ TEST_F(Dem_TEST, GPS)
   EXPECT_NEAR(sensor->Longitude().Degree(), longitude.Degree(), DOUBLE_TOL);
 
   // Sensor altitude is the elevation of the terrain + the sensor position.
-  EXPECT_NEAR(sensor->GetAltitude(),
+  EXPECT_NEAR(sensor->Altitude(),
       elevation + model->GetWorldPose().pos.z, 1);
 }
 #endif