+++ test/integration/physics_msgs.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2012-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -502,7 +502,7 @@ void PhysicsMsgsTest::SimpleShapeResize(const std::string &_physicsEngine)
     else
     {
       // Use physics API to resize
-      model->SetScale(scaleFactor * math::Vector3::One);
+      model->SetScale(scaleFactor * ignition::math::Vector3d::One);
     }
   }