+++ test/integration/surface_properties.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2015 Open Source Robotics Foundation
+ * Copyright (C) 2012-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -95,7 +95,7 @@ void SurfaceTest::CollideWithoutContact(const std::string &_physicsEngine)
     while (contacts.contact_size() == 0 && --steps > 0)
     {
       world->Step(1);
-      contacts = contactSensor->GetContacts();
+      contacts = contactSensor->Contacts();
     }
 
     // Verify that both objects are recognized by contact sensor
@@ -137,7 +137,7 @@ void SurfaceTest::CollideWithoutContact(const std::string &_physicsEngine)
     while (contacts.contact_size() == 0 && --steps > 0)
     {
       world->Step(1);
-      contacts = contactSensor->GetContacts();
+      contacts = contactSensor->Contacts();
     }
 
     // Verify that only contactBox is recognized by contact sensor